+++ packages/prelude/_src/Operations.ts
@@ -42,111 +42,45 @@ export class Operation extends MNModel<Operation, Operation.ConstructorInput, Op
 /** @ignore @internal @deprecated */
 export type OperationConstructor = typeof Operation
 
-//// codegen:start {preset: model}
-////
+// codegen:start {preset: model}
+//
 /* eslint-disable */
-export interface OperationProgress {
-  readonly completed: PositiveInt
-  readonly total: PositiveInt
-}
 export namespace OperationProgress {
   /**
    * @tsplus type OperationProgress.Encoded
+   * @tsplus companion OperationProgress.Encoded/Ops
    */
-  export interface Encoded {
-    readonly completed: number
-    readonly total: number
-  }
-  export const Encoded: EncodedOps = { $: {} }
-  // /**
-  //  * @tsplus type OperationProgress.Encoded/Aspects
-  //  */
-  // export interface EncodedAspects {}
-  /**
-   * @tsplus type OperationProgress.Encoded/Ops
-   */
-  export interface EncodedOps {}
+  export class Encoded extends EncodedClass<typeof OperationProgress>() {}
   export interface ConstructorInput
     extends ConstructorInputFromApi<typeof OperationProgress> {}
   export interface Props extends GetProvidedProps<typeof OperationProgress> {}
 }
-export interface Success {
-  readonly _tag: "Success"
-  readonly message: LongString | null
-}
 export namespace Success {
   /**
    * @tsplus type Success.Encoded
+   * @tsplus companion Success.Encoded/Ops
    */
-  export interface Encoded {
-    readonly _tag: "Success"
-    readonly message: string | null
-  }
-  export const Encoded: EncodedOps = { $: {} }
-  /**
-   * @tsplus type Success.Encoded/Aspects
-   */
-  export interface EncodedAspects {}
-  /**
-   * @tsplus type Success.Encoded/Ops
-   */
-  export interface EncodedOps { $: EncodedAspects }
+  export class Encoded extends EncodedClass<typeof Success>() {}
   export interface ConstructorInput
     extends ConstructorInputFromApi<typeof Success> {}
   export interface Props extends GetProvidedProps<typeof Success> {}
 }
-export interface Failure {
-  readonly _tag: "Failure"
-  readonly message: LongString | null
-}
 export namespace Failure {
   /**
    * @tsplus type Failure.Encoded
+   * @tsplus companion Failure.Encoded/Ops
    */
-  export interface Encoded {
-    readonly _tag: "Failure"
-    readonly message: string | null
-  }
-  export const Encoded: EncodedOps = { $: {} }
-  /**
-   * @tsplus type Failure.Encoded/Aspects
-   */
-  export interface EncodedAspects {}
-  /**
-   * @tsplus type Failure.Encoded/Ops
-   */
-  export interface EncodedOps { $: EncodedAspects }
+  export class Encoded extends EncodedClass<typeof Failure>() {}
   export interface ConstructorInput
     extends ConstructorInputFromApi<typeof Failure> {}
   export interface Props extends GetProvidedProps<typeof Failure> {}
 }
-export interface Operation {
-  readonly createdAt: Date
-  readonly id: StringId
-  readonly progress?: OperationProgress | undefined
-  readonly result?: Failure | Success | undefined
-  readonly updatedAt: Date | null
-}
 export namespace Operation {
   /**
    * @tsplus type Operation.Encoded
+   * @tsplus companion Operation.Encoded/Ops
    */
-  export interface Encoded {
-    readonly createdAt: string
-    readonly id: string
-    readonly progress?: OperationProgress.Encoded | undefined
-    readonly result?: Failure.Encoded | Success.Encoded | undefined
-    readonly updatedAt: string | null
-  }
-  export const Encoded: EncodedOps = { $: {} }
-  /**
-   * @tsplus type Operation.Encoded/Aspects
-   */
-  export interface EncodedAspects {}
-  /**
-   * @tsplus type Operation.Encoded/Ops
-   */
-  export interface EncodedOps { $: EncodedAspects }
+  export class Encoded extends EncodedClass<typeof Operation>() {}
   export interface ConstructorInput
     extends ConstructorInputFromApi<typeof Operation> {}
   export interface Props extends GetProvidedProps<typeof Operation> {}
@@ -154,4 +88,3 @@ export namespace Operation {
 /* eslint-enable */
 //
 // codegen:end
-//