+++ packages/vue/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @effect-app/vue
 
+## 0.23.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @effect-app/prelude@0.25.1
+  - @effect-app/vue@0.23.4
+
 ## 0.23.3
 
 ### Patch Changes